fix remove + get all

+++ src/main/java/fr/epsi/b32324c2/TestJpa.java
@@ -3,6 +3,7 @@ package fr.epsi.b32324c2;
 import fr.epsi.b32324c2.entites.Livre;
 
 import javax.persistence.*;
+import java.util.List;
 
 public class TestJpa {
 
@@ -10,12 +11,28 @@ public class TestJpa {
         EntityManagerFactory emf = Persistence.createEntityManagerFactory("pu-a1");
         EntityManager em = emf.createEntityManager();
 
-        Livre livre = new Livre(7);
-        em.remove(livre);
+        EntityTransaction et = em.getTransaction();
+        et.begin();
+        Livre livre = em.find(Livre.class, 7);
+        if (livre != null){
+            em.remove(livre);
+        }
+        et.commit();
 
+        // print all
+
+        Query queryAllBooks = em.createQuery("SELECT l FROM Livre l");
+        List<Livre> livres = queryAllBooks.getResultList();
+        System.out.println("Liste de tous les livres en base de données:");
+        for (Livre l : livres) {
+            System.out.println(l.toString());
+        }
 
         em.close();
         emf.close();
+
+
+
 
     }
 }
